ORCA-93 Renamed SDPSubarray folder name to ska_sdp_subarray and new version python package
