+++ src/tango_sdp_subarray/ska_sdp_subarray/register.py
