+++ src/tango_sdp_subarray/ska_sdp_subarray/schema/configure_pb.json
