+++ src/tango_sdp_subarray/tests/conftest.py
@@ -4,7 +4,7 @@
 import pytest
 from tango.test_context import DeviceTestContext
 
-from SDPSubarray import SDPSubarray
+from ska_sdp_subarray import SDPSubarray
 
 
 @pytest.fixture(scope='session', autouse=True)
