+++ src/tango_sdp_subarray/tests/test_subarray_device.py
@@ -12,7 +12,7 @@ from pytest_bdd import (given, parsers, scenarios, then, when)
 import tango
 from tango import DevState
 
-from SDPSubarray import AdminMode, ObsState
+from ska_sdp_subarray import AdminMode, ObsState
 
 
 # -----------------------------------------------------------------------------
